+++ rust/main/submitter/src/chain_tx_adapter/chains.rs
@@ -1,8 +1,10 @@
-mod builder;
+pub(crate) use factory::ChainTxAdapterFactory;
+pub(crate) use sealevel::SealevelPayload;
+pub(crate) use sealevel::SealevelTxPrecursor;
+
+mod factory;
+
+// chains modules below
 mod cosmos;
 mod ethereum;
 mod sealevel;
-
-pub(crate) use builder::ChainTxAdapterFactory;
-pub(crate) use sealevel::SealevelPayload;
-pub(crate) use sealevel::SealevelTxPrecursor;
